// src/models/users/AddUserModel.ts
export interface AddUserModel {
    
    firstName: string;
    lastName: string;
    userName: string;
    email: string;
    idNo: null | string;
    password: string;
    phoneNumber: string;
    birthDate?: string;
    cityId?: number;
    districtId?: number;
    address?: string;
    status?:boolean;
  }